Add `ensure => latest` when installing
Adds a `latest` option to the `ensure` parameter on `golang` and
`golang::installation`. It uses the `golang::latest_version()` function
to find the latest stable version of Go.

It also allows the version to be passed through `ensure`, like for
packages. This makes it possible to set `ensure => present` and just
leave any existing installation alone.

This deprecates `version` and `source` on `golang`.

This adds an `any_version` option to `ensure` on `golang::from_tarball`.
It indicates that the module should only install the tarball if there’s
no eisting installation, and that it should never upgrade it.