PDK update: release.rb --no-acceptance
Allows releasing when acceptance tests are broken (for reasons unrelated
to the module itself).

Submitting this as a PR so that acceptance tests will run in GitHub
Actions, since they don’t work on my laptop. If they pass, I’ll cut a Z
release to keep git and the released module more in sync.

This also adds a note to CHANGELOG.md.